+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # DGtalTools 0.9.3
 
+- *imageProcessing*:
+   - Creates imageProcessing directory. Add tools for doing image restoration and inpainting with Ambrosio-Tortorelli functional and discrete calculus. (Jacques-Olivier Lachaud, Marion Foare
+   [#280](https://github.com/DGtal-team/DGtalTools/pull/280))
+
 - *converters*:
    - fix tool itk2vol which was not able to read and convert int type image.
    (Bertrand Kerautret, [#276](https://github.com/DGtal-team/DGtalTools/pull/271))